Split plugin commands into individual files
